+++ R/pseudo_r2.R
@@ -5,9 +5,9 @@
 #'   including mixed and Bayesian regression models.
 #'
 #' @param x Fitted model of class \code{lm}, \code{glm}, \code{merMod},
-#'    \code{lme}, \code{plm}, \code{stanreg} or \code{brmsfit}. For method
-#'    \code{cod()}, only a \code{glm} or \code{glmer} with binrary response.
-#' @param n Optional, a \code{lmerMod} object, representing the fitted null-model
+#'    \code{glmmTMB}, \code{lme}, \code{plm}, \code{stanreg} or \code{brmsfit}.
+#'    For method \code{cod()}, only a \code{glm} with binrary response.
+#' @param n Optional, an \code{lme} object, representing the fitted null-model
 #'    (unconditional model) to \code{x}. If \code{n} is given, the pseudo-r-squared
 #'    for random intercept and random slope variances are computed
 #'    (\cite{Kwok et al. 2008}) as well as the Omega squared value
@@ -21,9 +21,8 @@
 #' @return For \code{r2()}, depending on the model, returns:
 #'         \itemize{
 #'           \item For linear models, the r-squared and adjusted r-squared values.
-#'           \item For linear mixed models, the r-squared and Omega-squared values.
+#'           \item For mixed models, the marginal and conditional r-squared values.
 #'           \item For \code{glm} objects, Cox & Snell's and Nagelkerke's pseudo r-squared values.
-#'           \item For \code{glmerMod} objects, Tjur's coefficient of determination.
 #'           \item For \code{brmsfit} or \code{stanreg} objects, the Bayesian version of r-squared is computed, calling \code{rstantools::bayes_R2()}.
 #'           \item If \code{loo = TRUE}, for \code{brmsfit} or \code{stanreg} objects a LOO-adjusted version of r-squared is returned.
 #'         }
@@ -36,6 +35,7 @@
 #'               \item Byrnes, J. 2008. Re: Coefficient of determination (R^2) when using lme() (\url{https://stat.ethz.ch/pipermail/r-sig-mixed-models/2008q2/000713.html})
 #'               \item Kwok OM, Underhill AT, Berry JW, Luo W, Elliott TR, Yoon M. 2008. Analyzing Longitudinal Data with Multilevel Models: An Example with Individuals Living with Lower Extremity Intra-Articular Fractures. Rehabilitation Psychology 53(3): 370–86. \doi{10.1037/a0012765}
 #'               \item Nakagawa S, Schielzeth H. 2013. A general and simple method for obtaining R2 from generalized linear mixed-effects models. Methods in Ecology and Evolution, 4(2):133–142. \doi{10.1111/j.2041-210x.2012.00261.x}
+#'               \item Nakagawa S, Johnson P, Schielzeth H (2017) The coefficient of determination R2 and intra-class correlation coefficient from generalized linear mixed-effects models revisted and expanded. J. R. Soc. Interface 14. \doi{10.1098/rsif.2017.0213}
 #'               \item Rabe-Hesketh S, Skrondal A. 2012. Multilevel and longitudinal modeling using Stata. 3rd ed. College Station, Tex: Stata Press Publication
 #'               \item Raudenbush SW, Bryk AS. 2002. Hierarchical linear models: applications and data analysis methods. 2nd ed. Thousand Oaks: Sage Publications
 #'               \item Snijders TAB, Bosker RJ. 2012. Multilevel analysis: an introduction to basic and advanced multilevel modeling. 2nd ed. Los Angeles: Sage
@@ -46,14 +46,18 @@
 #' @details For linear models, the r-squared and adjusted r-squared value is returned,
 #'          as provided by the \code{summary}-function.
 #'          \cr \cr
-#'          For linear mixed models, an r-squared approximation by computing the
+#'          For mixed models (from \pkg{lme4} or \pkg{glmmTMB}) marginal and
+#'          conditional r-squared values are calculated, based on
+#'          \cite{Nakagawa et al. 2017}.
+#'          \cr \cr
+#'          For \code{lme}-models, an r-squared approximation by computing the
 #'          correlation between the fitted and observed values, as suggested by
 #'          \cite{Byrnes (2008)}, is returned as well as a simplified version of
 #'          the Omega-squared value (1 - (residual variance / response variance),
 #'          \cite{Xu (2003)}, \cite{Nakagawa, Schielzeth 2013}), unless \code{n}
 #'          is specified.
 #'          \cr \cr
-#'          If \code{n} is given, for linear mixed models pseudo r-squared measures based
+#'          If \code{n} is given, for \code{lme}-models pseudo r-squared measures based
 #'          on the variances of random intercept (tau 00, between-group-variance)
 #'          and random slope (tau 11, random-slope-variance), as well as the
 #'          r-squared statistics as proposed by \cite{Snijders and Bosker 2012} and
@@ -63,22 +67,12 @@
 #'          For generalized linear models, Cox & Snell's and Nagelkerke's
 #'          pseudo r-squared values are returned.
 #'          \cr \cr
-#'          For generalized linear mixed models, the coefficient of determination
-#'          as suggested by \cite{Tjur (2009)} (see also \code{\link{cod}}). Note
-#'          that \emph{Tjur's D} is restricted to models with binary response.
-#'          \cr \cr
 #'          The ("unadjusted") r-squared value and its standard error for
 #'          \code{brmsfit} or \code{stanreg} objects are robust measures, i.e.
 #'          the median is used to compute r-squared, and the median absolute
 #'          deviation as the measure of variability. If \code{loo = TRUE},
 #'          a LOO-adjusted r-squared is calculated, which comes conceptionally
 #'          closer to an adjusted r-squared measure.
-#'          \cr \cr
-#'          More ways to compute coefficients of determination are shown
-#'          in this great \href{http://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#model-summaries-goodness-of-fit-decomposition-of-variance-etc.}{GLMM faq}.
-#'          Furthermore, see \code{\link[MuMIn]{r.squaredGLMM}} or
-#'          \code{\link[piecewiseSEM]{rsquared}} for conditional and marginal
-#'          r-squared values for GLMM's.
 #'
 #' @note \describe{
 #'         \item{\strong{cod()}}{
@@ -89,21 +83,21 @@
 #'          can be read like any other (Pseudo-)R-squared value.
 #'         }
 #'         \item{\strong{r2()}}{
-#'          If \code{n} is given, the Pseudo-R2 statistic is the proportion of
-#'          explained variance in the random effect after adding co-variates or
-#'          predictors to the model, or in short: the proportion of the explained
-#'          variance in the random effect of the full (conditional) model \code{x}
-#'          compared to the null (unconditional) model \code{n}.
+#'          For mixed models, the marginal r-squared considers only the variance
+#'          of the fixed effects, while the conditional r-squared takes both
+#'          the fixed and random effects into account.
+#'          \cr \cr
+#'          For \code{lme}-objects, if \code{n} is given, the Pseudo-R2 statistic
+#'          is the proportion of explained variance in the random effect after
+#'          adding co-variates or predictors to the model, or in short: the
+#'          proportion of the explained variance in the random effect of the
+#'          full (conditional) model \code{x} compared to the null (unconditional)
+#'          model \code{n}.
 #'          \cr \cr
 #'          The Omega-squared statistics, if \code{n} is given, is 1 - the proportion
 #'          of the residual variance of the full model compared to the null model's
 #'          residual variance, or in short: the the proportion of the residual
 #'          variation explained by the covariates.
-#'          \cr \cr
-#'          The r-squared statistics for linear mixed models, if the unconditional
-#'          model is also specified (see \code{n}), is the difference of the total
-#'          variance of the null and full model divided by the total variance of
-#'          the null model.
 #'          \cr \cr
 #'          Alternative ways to assess the "goodness-of-fit" is to compare the ICC
 #'          of the null model with the ICC of the full model (see \code{\link{icc}}).
@@ -132,11 +126,6 @@
 #' fit <- glm(services ~ neg_c_7 + c161sex + e42dep,
 #'            data = efc, family = binomial(link = "logit"))
 #' r2(fit)
-#'
-#' # Pseudo-R-squared values for random effect variances
-#' fit <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
-#' fit.null <- lmer(Reaction ~ 1 + (Days | Subject), sleepstudy)
-#' r2(fit, fit.null)
 #'
 #' @importFrom stats predict predict.glm residuals
 #' @export
@@ -179,10 +168,15 @@ r2 <- function(x, ...) {
 }
 
 
-#' @rdname cod
 #' @export
-r2.lmerMod <- function(x, n = NULL, ...) {
-  r2linmix(x, n)
+r2.lmerMod <- function(x, ...) {
+  r2_mixedmodel(x)
+}
+
+
+#' @export
+r2.glmmTMB <- function(x, ...) {
+  r2_mixedmodel(x)
 }
 
 
@@ -254,7 +248,8 @@ r2.glm <- function(x, ...) {
 
 #' @export
 r2.glmerMod <- function(x, ...) {
-  cod(x)
+  # cod(x)
+  r2_mixedmodel(x)
 }
 
 
@@ -383,4 +378,162 @@ r2linmix <- function(x, n) {
     # return results
     structure(class = "sj_r2", list(r2 = rsq, o2 = osq))
   }
+}
+
+
+#' @importFrom lme4 fixef getME VarCorr ranef findbars
+#' @importFrom stats family nobs var formula reformulate
+r2_mixedmodel <- function(x) {
+
+  ## Code taken from GitGub-Repo of package glmmTMB
+  ## Author: Ben Bolker, who used an
+  ## cleaned-up/adapted version of Jon Lefcheck's code from SEMfit
+
+  collapse_cond <- function(fit) {
+    if (is.list(fit) && "cond" %in% names(fit))
+      fit[["cond"]]
+    else
+      fit
+  }
+
+  faminfo <- model_family(x)
+
+  vals <- list(
+    beta = lme4::fixef(x),
+    X = lme4::getME(x, "X"),
+    vc = lme4::VarCorr(x),
+    re = lme4::ranef(x)
+  )
+
+  if (is(x,"glmmTMB")) {
+    vals <- lapply(vals, collapse_cond)
+
+    nullEnv <- function(x) {
+      environment(x) <- NULL
+      return(x)
+    }
+
+    if (!identical(nullEnv(x$modelInfo$allForm$ziformula), nullEnv(~0)))
+      warning("'r2()' ignores effects of zero-inflation.", call. = FALSE)
+
+    dform <- nullEnv(x$modelInfo$allForm$dispformula)
+
+    if (!identical(dform,nullEnv(~1)) && (!identical(dform,nullEnv(~0))))
+      warning("'r2()' ignores effects of dispersion model.", call. = FALSE)
+  }
+
+  ## Test for non-zero random effects
+  if (any(sapply(vals$vc, function(x) any(diag(x) == 0)))) {
+    ## TODO test more generally for singularity, via theta?
+    stop("Some variance components equal zero. Respecify random structure!")
+  }
+
+  ## Get variance of fixed effects: multiply coefs by design matrix
+  varF <- with(vals, stats::var(as.vector(beta %*% t(X))))
+
+  ## Are random slopes present as fixed effects? Warn.
+  random.slopes <- if ("list" %in% class(vals$re)) {
+    ## multiple RE
+    unique(c(sapply(vals$re,colnames)))
+  } else {
+    colnames(vals$re)
+  }
+
+  if (!all(random.slopes %in% names(vals$beta)))
+    warning("Random slopes not present as fixed effects. This artificially inflates the conditional R2. Respecify fixed structure!", call. = FALSE)
+
+  ## Separate observation variance from variance of random effects
+  nr <- sapply(vals$re, nrow)
+  not.obs.terms <- names(nr[nr != stats::nobs(x)])
+  obs.terms <- names(nr[nr == stats::nobs(x)])
+
+  ## Compute variance associated with a random-effects term
+  ## (Johnson 2014)
+  getVarRand <- function(terms) {
+    sum(sapply(
+      vals$vc[terms],
+      function(Sigma) {
+        Z <- vals$X[, rownames(Sigma), drop = FALSE]
+        Z.m <- Z %*% Sigma
+        return(sum(diag(crossprod(Z.m, Z))) / stats::nobs(x))
+      }))
+  }
+
+  ## Variance of random effects
+  varRand <- getVarRand(not.obs.terms)
+  sig <- attr(vals$vc, "sc")
+  if (is.null(sig)) sig <- 1
+
+  if (faminfo$is_linear) {
+    # get residual standard deviation sigma
+    varDist <- sig^2
+    varDisp <- 0
+  } else {
+    varDisp <- if (length(obs.terms) == 0 ) 0 else getVarRand(obs.terms)
+
+    badlink <- function(link, family) {
+      warning(sprintf("Model link '%s' is not yet supported for the %s distribution.", link, family), call. = FALSE)
+      return(NA)
+    }
+
+    if (faminfo$is_bin) {
+      varDist <- switch(
+        faminfo$link.fun,
+        logit = pi^2 / 3,
+        probit = 1,
+        badlink(faminfo$link.fun, faminfo$family)
+      )
+    } else if (faminfo$is_pois) {
+      ## Generate null model (intercept and random effects only, no fixed effects)
+
+      ## https://stat.ethz.ch/pipermail/r-sig-mixed-models/2014q4/023013.html
+      ## FIXME: deparse is a *little* dangerous
+      rterms <- paste0("(", sapply(lme4::findbars(stats::formula(x)), deparse), ")")
+      nullform <- stats::reformulate(rterms, response = ".")
+      null.model <- stats::update(x, nullform)
+
+      ## from MuMIn::rsquaredGLMM
+
+      ## Get the fixed effects of the null model
+      null.fixef <- unname(collapse_cond(lme4::fixef(null.model)))
+
+      ## in general want log(1+var(x)/mu^2)
+      logVarDist <- function(null.fixef) {
+        mu <- exp(null.fixef)
+        if (mu < 6)
+          warning(sprintf("mu of %0.1f is too close to zero, estimate may be unreliable \n", mu), call. = FALSE)
+
+        vv <- switch(
+          faminfo$family,
+          poisson = mu,
+          nbinom1 = ,
+          nbinom2 = stats::family(x)$variance(mu, sig),
+
+          if (is(x,"merMod"))
+            mu * (1 + mu / lme4::getME(x, "glmer.nb.theta"))
+          else
+            mu * (1 + mu / x$theta)
+        )
+
+        cvsquared <- vv / mu^2
+        return(log1p(cvsquared))
+      }
+
+      varDist <- switch(
+        faminfo$link.fun,
+        log = logVarDist(null.fixef),
+        sqrt = 0.25,
+        badlink(faminfo$link.fun, faminfo$family)
+      )
+    }
+  }
+
+  ## Calculate R2 values
+  rsq.marginal <- varF / (varF + varRand + varDisp + varDist)
+  rsq.conditional <- (varF + varRand) / (varF + varRand + varDisp + varDist)
+
+  names(rsq.marginal) <- "Marginal R2"
+  names(rsq.conditional) <- "Conditional R2"
+
+  structure(class = "sj_r2", list(rsq.marginal = rsq.marginal, rsq.conditional = rsq.conditional))
 }